refactor: #685 chat_service をオーケストレーション境界で分割する
ProcessChat を薄いオーケストレータにし、職種解決・完了判定・
質問生成・セッション要約をファイル分割する。公開 API は不変。
